Update Typings for Custom Instances of PM2
