+++ types/tsconfig.json
@@ -2,13 +2,10 @@
   "compilerOptions": {
       "module": "commonjs",
       "lib": ["es6"],
-      "noImplicitAny": true,
-      "noImplicitThis": true,
-      "strictNullChecks": true,
+      "strict": true,
 
       // If the library is an external module (uses `export`), this allows your test file to import "mylib" instead of "./index".
       // If the library is global (cannot be imported via `import` or `require`), leave this out.
-      "baseUrl": ".",
-      "paths": { "mylib": ["."] }
+      "baseUrl": "."
   }
 }
